ui improvements
Made lots of improvements to the UI so that it’s now a functional
prototype.  This includes a timer for rent and payroll payments (which
have been collapsed into a single timer of 60 seconds).

Bureaucrats are all now the same color unless the same color unless
they are a supervisor. Full color scheme can be turned back on by going
to the “Dev Mode” component in the PlayerController in the inspector,
and clicking on the checkbox for “Dev Mode Activated.”

The corruption will now start after the player has hired more than 16
nodes and more than 90 seconds has passes. Alternatively, you can just
hit space bar to start the corruption.